Update to add Yogurt ($YOG) to the token list
Yogurt is not yet another useless HRC-20 token on Harmony. But with a value, which gets defined by memes of yogurt and other milk based food.

It would be a huge step for the non-animal oriented meme token, to get listed in the official community token list of VenomDAO. Thanks!